Move beam convertion functions to separate module.
